feat(jobs): Groq enrichment fallback + shared prompt/schema across providers
Extract the enrichment prompt, JSON schema, and response parsing into
enrich-shared.ts so providers can't drift. Slim enrich-claude.ts to just the
Anthropic transport. Add enrich-groq.ts (gpt-oss-120b via the OpenAI-compatible
endpoint) and turn getEnrichmentProvider() into a Claude -> Groq -> heuristic
fallback chain: each step is gated on its key, and any error drops to the next
so enrichment never hard-fails.

Refine the relevance rule so full-time graduate/junior/entry-level roles at the
actual employer score normally (the cheaper model was wrongly rejecting genuine
early-career roles as "not a typical internship").

Add docs/opportunities-config.md — the gate-as-reviewer configuration runbook.
